+++ src/main/java/com/ias/ecommerce/infrastructure/persistence/repository/ProductRepositoryJpa.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepositoryJpa extends CrudRepository<ProductEntity, Integer> {
+public interface ProductRepositoryJpa extends CrudRepository<ProductEntity, Long> {
 
 }
